+++ R/difNLR-package.R
@@ -45,7 +45,7 @@
 #' Package: difNLR\cr
 #' Type: Package\cr
 #' Version: 1.4.2\cr
-#' Date: 2023-03-22\cr
+#' Date: 2023-05-02\cr
 #' Depends: R (>= 3.1)\cr
 #' Imports: calculus, ggplot2 (>= 3.4.0), msm, nnet, plyr, stats, VGAM\cr
 #' Suggests: ShinyItemAnalysis\cr