+++ R/estimNLR.R
@@ -132,7 +132,7 @@ estimNLR <- function(y, match, group, formula, method, lower, upper, start) {
     )
   )
   if (!is.null(M)) {
-    class(M) <- c("estNLR", class(M))
+    class(M) <- c("estimNLR", class(M))
   }
   return(M)
 }
@@ -150,7 +150,7 @@ estimNLR <- function(y, match, group, formula, method, lower, upper, start) {
   h. <- parse(text = as.character(formula)[3])
   h <- eval(h., envir = param)
 
-  l <- -sum((y * log(h)) + ((1 - y) * log(1 - h)), na.rm = T)
+  l <- -sum((y * log(h)) + ((1 - y) * log(1 - h)), na.rm = TRUE)
   return(l)
 }
 
@@ -206,9 +206,10 @@ residuals.lkl <- function(object, ...) {
   object$data$y - object$fitted
 }
 
-#' @rdname estNLR
+#' @rdname estimNLR
+#' @param x an object of \code{"estimNLR"} class.
 #' @export
-print.estNLR <- function(x, ...) {
+print.estimNLR <- function(x, ...) {
   formula <- switch(class(x)[2],
     "nls" = paste(x$m$formula()[2], x$m$formula()[1], x$m$formula()[3]),
     "lkl" = paste(x$formula[2], x$formula[1], x$formula[3]),
@@ -233,9 +234,14 @@ print.estNLR <- function(x, ...) {
   cat("\n", alg)
 }
 
-#' @rdname estNLR
+#' @rdname estimNLR
+#' @param object an object of \code{"estimNLR"} class.
+#' @param sandwich logical: should be sandwich estimator used for covariance
+#'   matrix of parameters when using \code{method = "nls"}? Default is
+#'   \code{FALSE}.
+#' @param ... other generic parameters for S3 methods.
 #' @export
-vcov.estNLR <- function(object, sandwich = FALSE, ...) {
+vcov.estimNLR <- function(object, sandwich = FALSE, ...) {
   if (inherits(object, "nls")) {
     if (sandwich) {
       e <- object$m$getEnv()